am57x6evm: configure cpusets for GPEVM variants
By deault only one core (0) is assigned for all categories of cpuset.
But we want both cores (0-1) to be assigned in case of AM57x.

Since we use same set of binaries for any AM57xEVM, configure the
cpusets dynamically using a script.

For AM572xEVM - the entries for foreground should be 0-1
For AM571xEVM - the entries for foreground should be 0

Signed-off-by: Vishal Mahaveer <vishalm@ti.com>
Signed-off-by: Praneeth Bajjuri <praneeth@ti.com>