Set 'media' user for i2c devices
The i2c devices are used by the radio and APPE user-space
components, mainly to configure the analog codecs for DSP
rendering and to configure radio tuners.

The user/group is set to media:system in order to allow
the APPE part to be started from the Audio HAL (which runs
as the 'media' user).

At the moment, not all i2c devices are accessed by the 'media'
user but having them all accessible by media user provides
more flexibility in case the analog codecs are moved to
different i2c buses.

Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>
[migrate from jacinto6evm to am57xevm]
Signed-off-by: Praneeth Bajjuri <praneeth@ti.com>

Change-Id: I397b8bd1becba40f78cc91d16838a062e6ee31d0