Set permissions for IPC daemon and device files
Add permissions and user/group for:
 * /dev/hwspinlock. It's used to lock/unlock specific hwspinlock
   instances through ioctls
 * /dev/uio0: It's is assumed to be user-space I/O driver for the
   SR0 shared region.
 * /dev/cmem

The user/group is set to media:system as the main users for these
IPC related devices will be audio (which runs as the 'media' user)
and radio (which is assumed to be a system app that runs in the
'system' group).

The LAD daemon user also needs to be changed to 'media' for the
same reasons described above.

Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>
[move from jacinto6evm to am57xevm]
Signed-off-by: Praneeth Bajjuri <praneeth@ti.com>

Change-Id: I10d955dfa9851050fab6f8eb6926599f4dfb9e8b