audio: Use XML audio configuration files
Replace the legacy audio_policy.conf with the new XML-based
audio policy configuration approach.

This version of the audio policy excludes the HDMI
HALs because they're currently not being loaded in the
binderized mode.

Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>
[picked from http://review.omapzoom.org/#/c/38586
and did not take vendor partition moving change as this is
not working with am57x yet]
Change-Id: Id2fd413a0a4af46ad769512edad9bb10bcf55edb
Signed-off-by: Praneeth Bajjuri <praneeth@ti.com>

Conflicts:
	device.mk